requirements.txt
requests
python-dotenv
pynetbox
requests
networkx
